requirements.txt
torch
torchvision
open-clip-torch
faiss-cpu
pandas
numpy
aiohttp
Pillow
scikit-learn
tqdm
